Switch controller to restcontroller annotation

+++ src/main/java/com/codecool/library/repository/BookDao.java
@@ -9,7 +9,7 @@ public interface BookDao {
 
     List<Book> getAllBooks();
 
-    int numberOfBooksInGenre(String genre);
+    long numberOfBooksInGenre(String genre);
 
     List<Book> listBooksByWritersBirthPlace(String birthPlace);
